Extract stats to separate package
